Assets, Events & Contexts: Exception Asset Consolidation (#1000)
Convert the 8 residual RaiseError.execute call sites (contexts/core.py,
contexts/feature.py x4, contexts/logging.py x2, blueprints/core.py) to a
new TiferetError.raise_error classmethod, then retire RaiseError,
events/static.py, and assets/exceptions.py entirely.

- assets/core.py: add a # *** classes section with TiferetError
  (gains raise_error classmethod) and TiferetAPIError (normalized
  __init__ signature: error_code, message=None, name=None, **kwargs,
  with name defaulting to error_code), moved from assets/exceptions.py.
- Delete assets/exceptions.py and events/static.py; relocate
  ParseParameter/ImportDependency into events/core.py unchanged.
- events/core.py::DomainEvent.raise_error now delegates to
  TiferetError.raise_error.
- assets/__init__.py, events/__init__.py: drop the RaiseError export.
- contexts/core.py, contexts/logging.py: replace the events-layer
  RaiseError/a passthrough import with direct assets imports.
- Update tests: relocate ParseParameter/ImportDependency coverage into
  tests/events/test_core.py, add TiferetError.raise_error/TiferetAPIError
  coverage to tests/assets/test_core.py, and fix stale
  tiferet.assets.exceptions / tiferet.events.static import references.
