+++ tests/events/test_core.py
@@ -2,12 +2,15 @@
 
 # *** imports
 
+# ** core
+import os
+
 # ** infra
 import pytest
 from unittest import mock
 
 # ** app
-from tiferet.events.core import DomainEvent, AsyncDomainEvent, TiferetError
+from tiferet.events.core import DomainEvent, AsyncDomainEvent, TiferetError, ParseParameter, ImportDependency, a
 
 # *** fixtures
 
@@ -689,3 +692,77 @@ async def test_handle_async_passthrough_sync_middleware():
     # The async runner awaits the returned coroutine, so the value passes through.
     assert result == 'result', 'Pass-through sync middleware should compose under handle_async'
     assert calls == ['pre', 'execute'], 'Sync pass-through middleware should run before execute'
+
+# ** test: test_parse_parameter_env_variable
+def test_parse_parameter_env_variable():
+    '''
+    Test that ParseParameter resolves an existing environment variable.
+    '''
+
+    # Set a test environment variable.
+    os.environ['TIFERET_TEST_VAR'] = 'hello_world'
+
+    # Parse the environment variable parameter.
+    result = ParseParameter.execute('$env.TIFERET_TEST_VAR')
+
+    # Verify the resolved value.
+    assert result == 'hello_world', 'Should resolve the environment variable value'
+
+    # Clean up.
+    del os.environ['TIFERET_TEST_VAR']
+
+# ** test: test_parse_parameter_missing_env_variable
+def test_parse_parameter_missing_env_variable():
+    '''
+    Test that ParseParameter raises on a missing environment variable.
+    '''
+
+    # Ensure the variable does not exist.
+    os.environ.pop('TIFERET_NONEXISTENT_VAR', None)
+
+    # Attempt to parse a missing env variable, expect TiferetError.
+    with pytest.raises(TiferetError) as exc_info:
+        ParseParameter.execute('$env.TIFERET_NONEXISTENT_VAR')
+
+    # Verify error code and kwargs.
+    assert exc_info.value.error_code == a.error.PARAMETER_PARSING_FAILED_ID, 'Should raise PARAMETER_PARSING_FAILED'
+    assert exc_info.value.kwargs.get('parameter') == '$env.TIFERET_NONEXISTENT_VAR', 'Should include the parameter'
+
+# ** test: test_parse_parameter_non_env_string
+def test_parse_parameter_non_env_string():
+    '''
+    Test that ParseParameter passes through a plain string unchanged.
+    '''
+
+    # Parse a non-environment variable string.
+    result = ParseParameter.execute('plain_value')
+
+    # Verify the value is returned unchanged.
+    assert result == 'plain_value', 'Should return the plain string unchanged'
+
+# ** test: test_import_dependency_success
+def test_import_dependency_success():
+    '''
+    Test that ImportDependency successfully imports a known class.
+    '''
+
+    # Import os.getenv via ImportDependency.
+    result = ImportDependency.execute('os', 'getenv')
+
+    # Verify the imported attribute is os.getenv.
+    assert result is os.getenv, 'Should import os.getenv successfully'
+
+# ** test: test_import_dependency_failure
+def test_import_dependency_failure():
+    '''
+    Test that ImportDependency raises on an invalid module/class.
+    '''
+
+    # Attempt to import a nonexistent module, expect TiferetError.
+    with pytest.raises(TiferetError) as exc_info:
+        ImportDependency.execute('nonexistent.module', 'FakeClass')
+
+    # Verify error code and kwargs.
+    assert exc_info.value.error_code == a.error.IMPORT_DEPENDENCY_FAILED_ID, 'Should raise IMPORT_DEPENDENCY_FAILED'
+    assert exc_info.value.kwargs.get('module_path') == 'nonexistent.module', 'Should include module_path'
+    assert exc_info.value.kwargs.get('class_name') == 'FakeClass', 'Should include class_name'
