+++ tiferet/assets/__init__.py
@@ -5,7 +5,6 @@
 __all__ = [
     'TiferetError',
     'TiferetAPIError',
-    'RaiseError',
     'ERROR_NOT_FOUND_ID',
     'DEFAULT_ERRORS',
     'core',
@@ -21,9 +20,9 @@ __all__ = [
 ]
 
 # ** app
-from .exceptions import TiferetError, TiferetAPIError, RaiseError
-from .error import ERROR_NOT_FOUND_ID, DEFAULT_ERRORS
+from .core import TiferetError, TiferetAPIError
 from . import core
+from .error import ERROR_NOT_FOUND_ID, DEFAULT_ERRORS
 from . import app
 from . import error
 from . import app
