+++ tiferet/events/__init__.py
@@ -9,9 +9,7 @@ __all__ = [
     'a',
     'ParseParameter',
     'ImportDependency',
-    'RaiseError',
 ]
 
 # ** app
-from .core import DomainEvent, AsyncDomainEvent, TiferetError, a
-from .static import ParseParameter, ImportDependency, RaiseError
+from .core import DomainEvent, AsyncDomainEvent, TiferetError, a, ParseParameter, ImportDependency
